Develop the code

+++ src/main/java/se/lexicon/model/VendingMachine.java
@@ -4,11 +4,11 @@ public interface VendingMachine {
 
     void addCurrency(int amount);
 
-    int getBalance();
+    double getBalance();
 
     Product request(int id);
 
-    int endSession();
+    double endSession();
 
     String getDescription(int id);
